fix remaining clippy lint

+++ lib/quantization/src/encoded_vectors_u8.rs
@@ -375,7 +375,7 @@ impl<TStorage: EncodedStorage> EncodedVectorsU8<TStorage> {
                     },
                 };
                 self.metadata
-                    .postprocess_score(score as f32, query.offset, vector_offset)
+                    .postprocess_score(score, query.offset, vector_offset)
             }
         }
     }
@@ -396,7 +396,7 @@ impl<TStorage: EncodedStorage> EncodedVectorsU8<TStorage> {
                     },
                 };
                 self.metadata
-                    .postprocess_internal_score(score as f32, query_offset, vector_offset)
+                    .postprocess_internal_score(score, query_offset, vector_offset)
             }
         }
     }
